requirements.txt
asgiref==3.9.1
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
pillow==11.3.0
sqlparse==0.5.3
